Script to generate cards using a template
Thanks for making Society available.

I used an online convertor to extract the text from the .pages files and
place it into a text file in the format "title \n <space> content \n\n".

With the SVG files you exported in #5 (thanks!), I created a SVG
template in templates/

The scripts in scripts/ will then produce a separate SVG file for each
of the cards in the text files. If the program "inkscape" is available,
a second script transforms the SVGs into PDFs and PNGs (also included).

The font used originally (Tsukushi) is only available without a license
on Mac OS X. I substituted with a webfont that looked compatible (Nunito)
which is also included.

If you want the files in a different folder, let me know and I can make
changes to the PR (the folders are hard-coded into the scripts).

I can run the script when there are new changes to the cards or I can
write instructions about how to run them under Mac OS X (with some
luck).